@@ -1,29 +1,28 @@
-﻿namespace Appy.Spatial.GeoJSON
-{
-    public abstract class Feature
-    {
-        public string Type { get; set; } = GeoType.Feature;
-        public string Id { get; set; }
-    }
+﻿namespace Appy.Spatial.GeoJSON;
 
-    public class Feature<TGeometry> : Feature where TGeometry : Geometry
-    {
-        public TGeometry Geometry { get; set; }
-    }
+public abstract class Feature
+{
+    public string Type { get; set; } = GeoType.Feature;
+    public string Id { get; set; }
+}
+
+public class Feature<TGeometry> : Feature where TGeometry : Geometry
+{
+    public TGeometry Geometry { get; set; }
+}
     
-    public class Feature<TGeometry, TProperties> : Feature<TGeometry> where TGeometry : Geometry
+public class Feature<TGeometry, TProperties> : Feature<TGeometry> where TGeometry : Geometry
+{
+    public Feature()
     {
-        public Feature()
-        {
-        }
-        
-        public Feature(string id, TGeometry geometry, TProperties properties)
-        {
-            Id = id;
-            Geometry = geometry;
-            Properties = properties;
-        }
-        
-        public TProperties Properties { get; set; }
     }
+        
+    public Feature(string id, TGeometry geometry, TProperties properties)
+    {
+        Id = id;
+        Geometry = geometry;
+        Properties = properties;
+    }
+        
+    public TProperties Properties { get; set; }
 }
