@@ -1,13 +1,12 @@
-﻿namespace Appy.Spatial.GeoJSON
+﻿namespace Appy.Spatial.GeoJSON;
+
+public class GeoType
 {
-    public class GeoType
-    {
-        public const string Feature = "Feature";
-        public const string Point = "Point";
-        public const string Polygon = "Polygon";
-        public const string MultiPolygon = "MultiPolygon";
-        public const string LineString = "LineString";
-        public const string MultiLineString = "MultiLineString";
-        public const string GeometryCollection = "GeometryCollection";
-    }
+    public const string Feature = "Feature";
+    public const string Point = "Point";
+    public const string Polygon = "Polygon";
+    public const string MultiPolygon = "MultiPolygon";
+    public const string LineString = "LineString";
+    public const string MultiLineString = "MultiLineString";
+    public const string GeometryCollection = "GeometryCollection";
 }
