@@ -1,21 +1,23 @@
-﻿namespace Appy.Spatial.GeoJSON
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Appy.Spatial.GeoJSON;
+
+public class MultiLineString : Geometry<IList<IList<IList<double>>>>
 {
-    public class MultiLineString : Geometry<IList<IList<IList<double>>>>
+    public MultiLineString()
+        : base(GeoType.MultiLineString)
     {
-        public MultiLineString() 
-            : base(GeoType.MultiLineString)
-        {
-        }
+    }
 
-        public MultiLineString(IList<IList<IList<double>>> coordinates) 
-            : base(GeoType.MultiLineString, coordinates)
-        {
-        }
+    public MultiLineString(IList<IList<IList<double>>> coordinates)
+        : base(GeoType.MultiLineString, coordinates)
+    {
+    }
 
-        public MultiLineString(IList<LineString> lineStrings) 
-            : base(GeoType.MultiLineString, lineStrings.Select(ls => ls.Coordinates).ToList())
-        {
-            
-        }
+    public MultiLineString(IList<LineString> lineStrings)
+        : base(GeoType.MultiLineString, lineStrings.Select(ls => ls.Coordinates).ToList())
+    {
+
     }
 }
