@@ -1,15 +1,16 @@
-﻿namespace Appy.Spatial.GeoJSON
-{
-    public class LineString : Geometry<IList<IList<double>>>
-    {
-        public LineString() 
-            : base(GeoType.LineString)
-        {
-        }
+﻿using System.Collections.Generic;
 
-        public LineString(IList<IList<double>> coordinates) 
-            : base(GeoType.LineString, coordinates)
-        {
-        }
+namespace Appy.Spatial.GeoJSON;
+
+public class LineString : Geometry<IList<IList<double>>>
+{
+    public LineString()
+        : base(GeoType.LineString)
+    {
+    }
+
+    public LineString(IList<IList<double>> coordinates)
+        : base(GeoType.LineString, coordinates)
+    {
     }
 }
