@@ -1,15 +1,16 @@
-﻿namespace Appy.Spatial.GeoJSON
-{
-    public class Polygon : Geometry<IList<IList<IList<double>>>>
-    {
-        public Polygon()
-            : base(GeoType.Polygon)
-        {
-        }
+﻿using System.Collections.Generic;
 
-        public Polygon(IList<IList<IList<double>>> coordinates)
-            : base(GeoType.Polygon, coordinates)
-        {
-        }
+namespace Appy.Spatial.GeoJSON;
+
+public class Polygon : Geometry<IList<IList<IList<double>>>>
+{
+    public Polygon()
+        : base(GeoType.Polygon)
+    {
+    }
+
+    public Polygon(IList<IList<IList<double>>> coordinates)
+        : base(GeoType.Polygon, coordinates)
+    {
     }
 }
